app/requirements.txt
flask==3.0.3
boto3==1.34.162
psycopg2-binary==2.9.9
gunicorn==21.2.0